refactoring large test files into smaller, more focused test files for better maintainability and clarity. This includes creating separate test files for base Bayesian regression functionality and specific linear regression tests, while removing redundant or outdated test files. Additionally, updating the base Bayesian regressor class to clarify the expected structure and responsibilities of subclasses, such as implementing specific methods for fitting and predicting with priors.
